MAINT: Move function optimal_parameters to separate module.
